requirements.txt
openai
python-pptx
requests
python-dotenv
